✏️ chore : 컬러 변수 red-01 추가
