+++ src/styles/_variables.scss
@@ -11,3 +11,5 @@ $gary-03: #f6f6f6;
 
 $white-01: #ffffff;
 $white-02: #f6f6f6;
+
+$red-01: #ed1c24;
